Window Cleaning (#63)
* Cleaned up a lot of the window code to make it simpler.
* Events are now stored in a vector which should be processed in the game loop, rather than via a callback. This just makes it far easier to see the order of operations in a single frame.
* Simplified the construction of events too, no more complex forwarding code; just pass the variant sub-elements directly. All event types just hold ints anyway so trivial to copy around.
* Rather than exposing separate `poll_events`, `swap_buffers` and `clear` functions, the window just has `begin_frame` and `end_frame` now, plus an `events` functions to return a span of the events for the given frame.
* Removed a bunch of unused functions from the window.
* Removed the mouse pos stored in the variable, you should get this from the `mouse` object instead.
* Move a bunch of editor specific values to the editor struct.
* Added some code to load files, currently doesn't do anything except print the size, but in the future I want to make it possible to load PNG files to construct worlds from. I'll expand on this with a proper editor.

+++ src/camera.hpp
@@ -6,8 +6,8 @@ namespace sand {
 struct camera
 {
     glm::vec2 top_left;
-    float screen_width;
-    float screen_height;
+    int screen_width;
+    int screen_height;
     float world_to_screen; // the number of screen pixels per world pixel, multiplying by it takes you from world space to screen space
 };
